Reversed flow of input in NetworkTurnEntity, no need for it to be abstract as both extending classes would be exactly the same.
